// BeltmanSoftwareDesign.Angular/BeltmanSoftwareDesign.Angular/src/app/interfaces/company.ts
export interface Company {
    id: number;
    countryId: number | null;
    countryName: string | null;
    name: string | null;
    address: string | null;
    postalcode: string | null;
    place: string | null;
    phoneNumber: string | null;
    email: string | null;
    website: string | null;
    btwNumber: string | null;
    kvkNumber: string | null;
    iban: string | null;
}